controller tests added successfully

+++ src/app.service.ts
@@ -4,18 +4,14 @@ import { Model } from 'mongoose';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User, UserDocument } from './schema/user.schema';
+import { UserAlreadyExists } from './user-already-exists.exception';
+import { UserDontExists } from './user-not-exist';
 
 @Injectable()
 export class AppService {
   constructor(
     @InjectModel(User.name) private readonly model: Model<UserDocument>,
   ) {}
-  // getHello(): string {
-  //   return 'Hello World!';
-  // }
-  // getAnotherHello(): string {
-  //   return `Hello World new !`;
-  // }
   async findAll(): Promise<User[]> {
     return await this.model.find().exec();
   }
@@ -23,14 +19,23 @@ export class AppService {
     return await this.model.findById(id).exec();
   }
   async create(createUserDto: CreateUserDto): Promise<User> {
-    return await new this.model({
-      ...createUserDto,
-    }).save();
+    const existingUser = await this.model.findOne({
+      email: createUserDto.email,
+    });
+    if (existingUser) {
+      throw new UserAlreadyExists();
+    }
+    const createdUser = new this.model(createUserDto);
+    return await createdUser.save();
   }
   async update(id: string, updateUserDto: UpdateUserDto): Promise<User> {
     return await this.model.findByIdAndUpdate(id, updateUserDto).exec();
   }
   async delete(id: string): Promise<User> {
+    const notExistingUser = await this.model.findById(id);
+    if (!notExistingUser) {
+      throw new UserDontExists();
+    }
     return await this.model.findByIdAndDelete(id).exec();
   }
 }
